Finished Task
I found a way to keep a reference to a leader within a KanbanModel. I created a wrapper for the KanbanModel class and put that extra property in there. It was very important I try not to display that property onto the KanbanModel card when inside the SfKanban because then the default Syncfusion adapter would no longer work properly. Therefore I would loose a multitude of the cool functionalities.