Add Scroll View / Started Delete Boards
Added a scroll view as a main container for both EditCardActivity and EditBoardActivity. This allows the user to scroll up when the soft keyboard is toggled and covers any edit texts. For this to work properly I needed to implement an XML property for the listviews to be able to scroll inside another scroll view. Also start implementing the UI required for deleting boards. With deleting boards I've also updated the create boards button.